refactor: move testCleanup near the AfterEach function
