+++ tests/utils.go
@@ -106,16 +106,6 @@ const (
 	defaultDiskSize        = "1Gi"
 )
 
-func TestCleanup() {
-	GinkgoWriter.Println("Global test cleanup started.")
-	testsuite.CleanNamespaces()
-	libnode.CleanNodes()
-	resetToDefaultConfig()
-	testsuite.EnsureKubevirtReady()
-	SetupAlpineHostPath()
-	GinkgoWriter.Println("Global test cleanup ended.")
-}
-
 func SetupAlpineHostPath() {
 	const osAlpineHostPath = "alpine-host-path"
 	libstorage.CreateHostPathPv(osAlpineHostPath, testsuite.GetTestNamespace(nil), testsuite.HostPathAlpine)
@@ -1586,19 +1576,6 @@ func UpdateKubeVirtConfigValueAndWait(kvConfig v1.KubeVirtConfiguration) *v1.Kub
 	log.DefaultLogger().Infof("system is in sync with kubevirt config resource version %s", kv.ResourceVersion)
 
 	return kv
-}
-
-// resetToDefaultConfig resets the config to the state found when the test suite started. It will wait for the config to
-// be propagated to all components before it returns. It will only update the configuration and wait for it to be
-// propagated if the current config in use does not match the original one.
-func resetToDefaultConfig() {
-	if !CurrentSpecReport().IsSerial {
-		// Tests which alter the global kubevirt config must be run serial, therefor, if we run in parallel
-		// we can just skip the restore step.
-		return
-	}
-
-	UpdateKubeVirtConfigValueAndWait(testsuite.KubeVirtDefaultConfig)
 }
 
 type compare func(string, string) bool
